requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
kombu==5.3.2
magic-filter==1.0.11
multidict==6.0.4
playwright==1.38.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
pyee==9.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.0
six==1.16.0
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
